WIP: Manually enable the I8MM, SVE and SVE2 cpuflags on Windows
We could detect SVE and SVE2, and I8MM via SVE-I8MM, but it
requires using a new enough version of Wine (which is available)
and new enough versions of llvm-mingw (which we install) and
MSVC/WinSDK (which doesn't get installed yet) to expose those feature
constants.

As we're running tests in QEMU, we want to test all the possible
codepaths, to verify that the code has been assembled correctly.